Fix existing programmed transform errors
