+++ packages/mdd-engine/src/adapted-programs/programmable-units/engine-program-model/item-definition/parseItemDefinition.ts
@@ -19,6 +19,9 @@ import {
   ItemDefinitionModelStreamMetatype,
 } from './itemDefinitionModel';
 
+/**
+ * Parses the type or class definition of a streamable item
+ */
 export const parseItemDefinition = buildProgrammedTransform({
   name: 'parseItemDefinition',
 })
